implement the SitemapLoader class
