+++ docs/api/com/kazurayam/inspectus/selenium/WebDriverFormulas.html
@@ -98,7 +98,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>com.kazurayam.inspectus.selenium.WebDriverFormulas</li>
+<li>selenium.com.kazurayam.inspectus.materialize.WebDriverFormulas</li>
 </ul>
 </li>
 </ul>
