+++ docs/index.md
@@ -6,7 +6,7 @@ Back to the [project’s GitHub repository](https://github.com/kazurayam/inspect
 
 The inspectus4selenium provides a small Java class
 
--   [com.kazurayam.inspectus.selenium.WebDriverFormulas](https://github.com/kazurayam/inspectus4selenium/blob/master/src/main/java/com/kazurayam/inspectus/selenium/WebDriverFormulas.java)
+-   [selenium.com.kazurayam.inspectus.materialize.WebDriverFormulas](https://github.com/kazurayam/inspectus4selenium/blob/master/src/main/java/com/kazurayam/inspectus/selenium/WebDriverFormulas.java)
 
 When you write a web UI test on top of [Selenium WebDriver](https://www.selenium.dev/documentation/webdriver/), you need to be aware of timing issues: you have to repeat writing codes so that your tests wait for pages to be loaded into the browser completely; you have to wait for web element to be present in the DOM before doing any action on it. You will write codes to wait for conditions to meet repeatedly. Your UI tests tend to be verbose and difficult to maintain.
