+++ settings.gradle
@@ -6,6 +6,9 @@ dependencyResolutionManagement {
             version('ashot', '1.5.4')
             library('ashot', 'ru.yandex.qatools.ashot', 'ashot').versionRef('ashot')
 
+            version('commons-csv', '1.9.0')
+            library('commons-csv', 'org.apache.commons','commons-csv').versionRef('commons-csv')
+
             version('httpclient', '5.2')
             library('httpclient', 'org.apache.httpcomponents.client5', 'httpclient5').versionRef('httpclient')
