+++ src/main/java/com/kazurayam/inspectus/materialize/selenium/WebPageMaterializingFunction.java
@@ -1,0 +1,9 @@
+package com.kazurayam.inspectus.materialize.selenium;
+
+import com.kazurayam.materialstore.core.filesystem.MaterialstoreException;
+
+@FunctionalInterface
+public interface WebPageMaterializingFunction<WebDriver, Target, Map, Material> {
+
+    Material accept(WebDriver driver, Target target, Map attributes) throws MaterialstoreException;
+}
