+++ src/main/java/com/kazurayam/inspectus/materialize/url/URLMaterializingFunction.java
@@ -1,0 +1,8 @@
+package com.kazurayam.inspectus.materialize.url;
+
+import com.kazurayam.materialstore.core.filesystem.MaterialstoreException;
+
+public interface URLMaterializingFunction<Target, Material> {
+
+    Material accept(Target target) throws MaterialstoreException;
+}
