--- src/main/java/com/kazurayam/inspectus/selenium/WebPageMaterializingFunction.java
@@ -1,9 +1,0 @@
-package com.kazurayam.inspectus.selenium;
-
-import com.kazurayam.materialstore.core.filesystem.MaterialstoreException;
-
-@FunctionalInterface
-public interface WebPageMaterializingFunction<WebDriver, Target, Map, Material> {
-
-    Material accept(WebDriver driver, Target target, Map attributes) throws MaterialstoreException;
-}
