--- src/main/java/com/kazurayam/inspectus/url/URLMaterializingFunction.java
@@ -1,8 +1,0 @@
-package com.kazurayam.inspectus.url;
-
-import com.kazurayam.materialstore.core.filesystem.MaterialstoreException;
-
-public interface URLMaterializingFunction<Target, Material> {
-
-    Material accept(Target target) throws MaterialstoreException;
-}
