+++ src/test/fixtures/com/kazurayam/inspectus/materialize/discovery/ReadingCSVTest/names.csv
@@ -1,0 +1,2 @@
+"First Name","Last Name"
+"暴れん坊将軍 徳川吉宗","松平 健"
